requirements.txt
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pychalk==2.0.1
pycodestyle==2.5.0
pylint==2.3.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
typed-ast==1.3.5
urllib3==1.25.2
virtualenv==16.6.0
wrapt==1.11.1
